+++ MOP/src/RuleFiles/Cases/IgnoreRule.cs
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.If not, see<http://www.gnu.org/licenses/>.
 
-using UnityEngine;
-
 namespace MOP
 {
     class IgnoreRule
@@ -51,36 +49,6 @@ namespace MOP
         {
             this.MainObject = MainObject;
             this.ObjectName = ObjectName;
-        }
-    }
-
-    // This class is intended for special flags used in specific cases.
-    class SpecialRules
-    {
-        public bool SatsumaIgnoreRenderers;
-        public bool DontDestroyEmptyBeerBottles;
-        
-        // Custom rule files.
-        public bool IgnoreModVehicles;
-        public bool ToggleAllVehiclesPhysicsOnly;
-        public bool DrivewaySector;
-        public bool ExperimentalSatsumaTrunk;
-        public bool ExperimentalOptimization;
-    }
-
-    class NewSector
-    {
-        public Vector3 Position;
-        public Vector3 Rotation;
-        public Vector3 Scale;
-        public string[] Whitelist;
-
-        public NewSector(Vector3 Position, Vector3 Scale, Vector3 Rotation, string[] Whitelist)
-        {
-            this.Position = Position;
-            this.Scale = Scale;
-            this.Rotation = Rotation;
-            this.Whitelist = Whitelist;
         }
     }
 }