+++ MOP/src/WorldManager.cs
@@ -93,6 +93,9 @@ namespace MOP
             // Looking for player and yard
             player = GameObject.Find("PLAYER").transform;
 
+            // Add GameFixes MonoBehaviour.
+            gameObject.AddComponent<GameFixes>();
+
             // Loading vehicles
             vehicles = new List<Vehicle>
             {
@@ -960,111 +963,9 @@ namespace MOP
             }
         }
 
-        public void HoodFix(Transform hoodPivot, Transform batteryPivot, Transform batteryTrigger)
+        public Vehicle GetFlatbed()
         {
-            StartCoroutine(HoodFixCoroutine(hoodPivot, batteryPivot, batteryTrigger));
-        }
-
-        IEnumerator HoodFixCoroutine(Transform hoodPivot, Transform batteryPivot, Transform batteryTrigger)
-        {
-            yield return new WaitForSeconds(1);
-
-            // Hood
-            Transform hood = GameObject.Find("hood(Clone)").transform;
-            CustomPlayMakerFixedUpdate hoodFixedUpdate = hood.gameObject.AddComponent<CustomPlayMakerFixedUpdate>();
-
-            // Fiber Hood
-            GameObject fiberHood = Resources.FindObjectsOfTypeAll<GameObject>()
-                .First(obj => obj.name == "fiberglass hood(Clone)"
-                && obj.GetComponent<PlayMakerFSM>() != null
-                && obj.GetComponent<MeshCollider>() != null);
-
-            int retries = 0;
-            if (MopFsmManager.IsStockHoodBolted() && hood.parent != hoodPivot)
-            {
-                while (hood.parent != hoodPivot)
-                {
-                    // Satsuma got disabled while trying to fix the hood.
-                    // Attempt to fix it later.
-                    if (!hoodPivot.gameObject.activeSelf)
-                    {
-                        Satsuma.instance.AfterFirstEnable = false;
-                        yield break;
-                    }
-
-                    MopFsmManager.ForceHoodAssemble();
-                    yield return null;
-
-                    // If 10 retries failed, quit the loop.
-                    retries++;
-                    if (retries == 10)
-                    {
-                        break;
-                    }
-                }
-            }
-
-            hoodFixedUpdate.StartFixedUpdate();
-
-            if (fiberHood != null && MopFsmManager.IsFiberHoodBolted() && fiberHood.transform.parent != hoodPivot)
-            {
-                retries = 0;
-                while (fiberHood.transform.parent != hoodPivot)
-                {
-                    // Satsuma got disabled while trying to fix the hood.
-                    // Attempt to fix it later.
-                    if (!hoodPivot.gameObject.activeSelf)
-                    {
-                        Satsuma.instance.AfterFirstEnable = false;
-                        yield break;
-                    }
-
-                    MopFsmManager.ForceHoodAssemble();
-                    yield return null;
-
-                    // If 10 retries failed, quit the loop.
-                    retries++;
-                    if (retries == 60)
-                    {
-                        break;
-                    }
-                }
-            }
-
-            hood.gameObject.AddComponent<SatsumaBoltsAntiReload>();
-            fiberHood.gameObject.AddComponent<SatsumaBoltsAntiReload>();
-
-            // Adds delayed initialization for hood hinge.
-            if (hood.gameObject.GetComponent<DelayedHingeManager>() == null)
-                hood.gameObject.AddComponent<DelayedHingeManager>();
-
-            // Fix for hood not being able to be closed.
-            if (hood.gameObject.GetComponent<SatsumaCustomHoodUse>() == null)
-                hood.gameObject.AddComponent<SatsumaCustomHoodUse>();
-
-            // Fix for battery popping out.
-            if (MopFsmManager.IsBatteryInstalled() && batteryPivot.parent == null)
-            {
-                batteryTrigger.gameObject.SetActive(true);
-                batteryTrigger.gameObject.GetComponent<PlayMakerFSM>().SendEvent("ASSEMBLE");
-            }
-        }
-
-        public void KekmetTrailerAttach()
-        {
-            StartCoroutine(KekmetTrailerAttachCoroutine());
-        }
-
-        IEnumerator KekmetTrailerAttachCoroutine()
-        {
-            while (!vehicles[6].gameObject.activeSelf)
-            {
-                yield return new WaitForSeconds(1f);
-            }
-
-            vehicles[6].transform.rotation = vehicles[6].Rotation;
-            vehicles[6].transform.position = vehicles[6].Position;
-            PlayMakerFSM.BroadcastEvent("TRAILERATTACH");
+            return vehicles[6];
         }
     }
 }